feat: check Node version

+++ 04-copy-directory/index.js
@@ -4,6 +4,13 @@ const DIR = './04-copy-directory/'
 const DIR_SOURCE = 'files'
 const DIR_TARGET = 'files-copy'
 
+let version =  process.version.match(/v(\d+)\.(\d+)\.(\d+)/);
+if(version[1]<16 || version[2]<13){
+    console.log("Вы запускаете скрип под Node "+process.version);
+    console.log("Проверка проводится на Node 16.13.0 LTS");
+    process.exit();
+}
+
 new Promise((res, rej) => {
     fs.rm(DIR + DIR_TARGET, { recursive: true }, (e, r) => {
         // if (e) rej(e);
